# app-service/src/main/java/es/codeurjc/practica2/dto/ReviewDTO.java
package es.codeurjc.practica2.dto;

public record ReviewDTO(
        Long id,
        String comment,
        Integer rating,
        Long userId,
        String userName,
        Long bookId,
        String bookTitle
) {
}